Removed health endpoint dependency from startup probe.
I initially used the health endpoint for convenience, but really the startup
probe just needs to verify name resolution works. jshell is nice because it's
preinstalled and I can perform the lookup in the same way hedera-service does.

The first time jshell is run it generates output on stderr. I ran jshell twice
and ignored stderr the firt time to suppress this output, but preserve genuine
errors the second time.

The cluster domain is hard coded as "cluster.local" but really it should be
configurable.

I tested this by using solo to bring up and start nodes, then delete the
namespace and repeat. I found that this could fail, meaning the startup probe
would succeed but then the hedera service would fail. I found adding a delay
after tearing down the environment and before bringing up a new environment
succeeded reliably. Presumably the startup probe saw an old DNS record before
it expired and hedera service was not so "lucky".

Signed-off-by: Matt Riben <matt.riben@swirldslabs.com>